requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==21.1
djangorestframework==3.12.4
mssql-django==1.1.1
PyJWT==2.1.0
pyodbc==4.0.32
pytz==2021.3
rest-framework-generic-relations==2.1.0
sqlparse==0.4.2
whitenoise==5.3.0
